+++ Ratios/Helpers/PersistenceStore.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 protocol PersistenceStore {
-    func save(grounds: Double, ratio: Int)
-    func getValues() -> (grounds: Double, ratio: Int)
+    func save(grounds: Double, ratio: Double)
+    func getValues() -> (grounds: Double, ratio: Double)
 }
 
 extension UserDefaults: PersistenceStore {
@@ -18,14 +18,14 @@ extension UserDefaults: PersistenceStore {
     private static let ratiosPersistenceGroundsKey = "co.brushedtype.ratios.grounds"
     private static let ratiosPersistenceRatioKey = "co.brushedtype.ratios.ratio"
 
-    func save(grounds: Double, ratio: Int) {
+    func save(grounds: Double, ratio: Double) {
         self.set(grounds, forKey: UserDefaults.ratiosPersistenceGroundsKey)
         self.set(ratio, forKey: UserDefaults.ratiosPersistenceRatioKey)
     }
 
-    func getValues() -> (grounds: Double, ratio: Int) {
+    func getValues() -> (grounds: Double, ratio: Double) {
         let grounds = self.double(forKey: UserDefaults.ratiosPersistenceGroundsKey)
-        let ratio = self.integer(forKey: UserDefaults.ratiosPersistenceRatioKey)
+        let ratio = self.double(forKey: UserDefaults.ratiosPersistenceRatioKey)
 
         let groundsOrFallback = grounds > 0 ? grounds : 12
         let ratioOrFallback = ratio > 0 ? ratio : 16
